Rename module to gridded_models
