+++ photutils/psf/__init__.py
@@ -4,12 +4,12 @@ This subpackage contains tools to perform point-spread-function (PSF)
 photometry.
 """
 
-from . import (epsf, epsf_stars, functional_models, griddedpsfmodel, groupers,
+from . import (epsf, epsf_stars, functional_models, gridded_models, groupers,
                image_models, photometry, utils)
 from .epsf import *  # noqa: F401, F403
 from .epsf_stars import *  # noqa: F401, F403
 from .functional_models import *  # noqa: F401, F403
-from .griddedpsfmodel import *  # noqa: F401, F403
+from .gridded_models import *  # noqa: F401, F403
 from .groupers import *  # noqa: F401, F403
 from .image_models import *  # noqa: F401, F403
 from .matching import *  # noqa: F401, F403
@@ -21,7 +21,7 @@ __all__ = []
 __all__ += epsf.__all__
 __all__ += epsf_stars.__all__
 __all__ += functional_models.__all__
-__all__ += griddedpsfmodel.__all__
+__all__ += gridded_models.__all__
 __all__ += groupers.__all__
 __all__ += image_models.__all__
 __all__ += photometry.__all__
